requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pymongo==4.10.1
numpy==1.26.4
python-dotenv==1.0.1
pytest==8.3.3
requests==2.32.3
httpx==0.28.0
tensorflow==2.18.0
tf-keras==2.18.0
deepface==0.0.93
ultralytics==8.3.39
